Add all MatLab Scripts and folders from ModEM SVN Repo
This commit adds all matlab files and folders found in the ModEM SVN repo,
mainly the trunk/matlab folder as well as what was found in the ModelPlot
repository.

It may contain more then what we want to release, but at least everything will
be recorded in the GitHub repository. We should take some time to whittle
things down to 'supported' and 'unsupported' (i.e. things that work and things
that don't work).