Large update - enable ability to read/write grid and covariance files
This update updates ModEMGrid the ability to accurately read and write files.
Before this was possible, but this adds some cleanup and ensures things are
writing out correctly. There hasn't been much testing/confirmation before these
updates, but now they do.

A lot of the reading and writing of rhos comes from mtpy.

This commit also updates the ability to create a covariance file given a modem
rho file. It adds a script which demonstrates this. So far, it has only been
testing by adding one mask to water.