Update siteChar creation in readApres_3D to match that of readZ_3D
The creation of siteChar was updated in readZ_3D. This update turned siteLoc
into a cell array, rather than a char array.

readApres_3D was still able to read, but calling writeZ_3D (with the APRES flag
set to True), resulted in an error as writeZ_3D was expecting a cell array.

This commit updates readApres_3D to also create a cell array for it's siteChar.

While this commit fixes the siteChar issue, there are still issues with writing
out data that is read in from readApres_3D via writeZ_3D, which will be
addressed by a future commit.