Use an integer rather than hard coded numbers when reading in data
Previous to this commit readZ_3D expects that Full_Impedance is the first data
type read and that Full_Vertical_Components is always read second.

This change allows more flexibility in the data types and allows for
Full_Vertical_Components to be in the data file either before or after.

Likewise, it should allow there to be any number of different data types, but
this needs to be tested further.