Update 'onetype' in readZ_3D and readApres_3D
This change updates the code in readZ_3D and readApres_3D to better handle the
`oneType` argument.

In readZ_3D, it fixes a problem when it was resulting in an error. It also adds
an error message if a requested data type is not present, and displays the
datatypes found.

Before this commit in readZ_3D with `onetype`, the return argument
`units` would be that of the *last* datatype read. This change updates it so
the units associated with `onetype` are returned.

Similar fixes are done in readApres_3D; however, it currently doesn't produce a
similar error message.