Fix up readApres_3D, readZ/writeZ_3D to work with more data types
Before this commit there were issues with readApres_3D, readZ_3D and writeZ_3D
that produced errors when writing out certain data types.

Specifically, mostly Full_Impedance, or Full_Impedance followed by
Full_Vertical_Components could be read. Other data types could also be read,
but when they were read, they would be associated with the incorrect
components.

This is because writeZ_3D made assumptions on the datatype, and the components,
based on the number of components. This limited writeZ to only
Full_Vertical_Components, Full_Impedance, or Full_Impedance followed by
Vertical_Components.

These changes update readApres_3D, readZ_3D and writeZ_3D so that we can write
out other data types. To do this, when we read in the data, we assign the type
to both the allData and the info structures that the reads return.

This is similar to what is done in mtdata, where the type is associated with
the object.

The allData struct that is returned by the reads includes *all* the data of all
read components of a file (in data{period}.Z). In this case, both data types
are now included as a strings array in the order that they appear. However, if
Full_Impedance and Full_Vertical_Components are read, they will always be
returned in that order, to ensure compatibility with other scripts.

The writes then use this information to determine how to write out data.
Because `data{period}.Z` could contain any number of different data types the
write scripts use the types in the allData structure to determine the order of
this data. (For instance, write out Full_Impedance components, then
Off_Diagonal_impedance components).

If the types are not present, writeZ tries to determine the datatype by the
listed components.

These changes have worked to make the reads and writes a bit more optimized,
while maintaining the original interface to ensure other scripts aren't broken.

These changes also should allow for the scripts to be able to read and write
any number of datatypes that are found with in a file.